move calorie and weight tracker server REST api to util files for reusability
